Move background colour auto-detect script to where the other webpack files are
