+++ app/javascript/mastodon/svg_select.js
